requirements.txt
git+https://github.com/AlexanderMoreR/fbchat.git
flask
python-dotenv
selenium